backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
google-genai
bcrypt
pyjwt
werkzeug
groq